One method remaining

+++ src/main/java/vm/mtgdeckrepository/service/impl/DeckServiceJpa.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 import vm.mtgdeckrepository.dao.CardInDeckRepository;
 import vm.mtgdeckrepository.dao.DeckRepository;
-import vm.mtgdeckrepository.dao.PlayerRepository;
 import vm.mtgdeckrepository.domain.CardInDeck;
 import vm.mtgdeckrepository.domain.Deck;
 import vm.mtgdeckrepository.domain.Player;
@@ -13,8 +12,10 @@ import vm.mtgdeckrepository.dto.CardInDeckDTO;
 import vm.mtgdeckrepository.error.EntityMissingException;
 import vm.mtgdeckrepository.service.DeckService;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 @Service
 public class DeckServiceJpa implements DeckService {
@@ -93,17 +94,25 @@ public class DeckServiceJpa implements DeckService {
 
         Deck newDeck = deckRepo.save(deck);
 
+        Set<CardInDeck> cards = new HashSet<>();
+
         for (CardInDeckDTO card : main_deck) {
-            cardInDeckRepo.save(new CardInDeck(newDeck, card.getCard_name(), card.getQuantity(), true));
+            CardInDeck newCard = new CardInDeck(newDeck, card.getCard_name(), card.getQuantity(), true);
+            cardInDeckRepo.save(newCard);
+            cards.add(newCard);
         }
 
         for (CardInDeckDTO card : sideboard) {
-            cardInDeckRepo.save(new CardInDeck(newDeck, card.getCard_name(), card.getQuantity(), false));
+            CardInDeck newCard = new CardInDeck(newDeck, card.getCard_name(), card.getQuantity(), false);
+            cardInDeckRepo.save(newCard);
+            cards.add(newCard);
         }
 
-        Deck newNewDeck = deckRepo.findById(newDeck.getDeck_id()).get();
+        newDeck.setCards(cards);
 
-        return newNewDeck;
+        deckRepo.save(newDeck);
+
+        return newDeck;
     }
 
     @Override
@@ -125,6 +134,17 @@ public class DeckServiceJpa implements DeckService {
         Optional<Deck> deckCheck = deckRepo.findById(id);
         if (deckCheck.isEmpty()) {
             throw new EntityMissingException("Deck not found.");
+        }
+
+        String[] legalFormats = {"Standard", "Pioneer", "Modern", "Pauper"};
+        boolean correctFormat = false;
+        for (String legalFormat : legalFormats) {
+            if (legalFormat.equals(format)) {
+                correctFormat = true;
+            }
+        }
+        if (!correctFormat) {
+            throw new IllegalArgumentException("Format not supported.");
         }
 
         String[] basicLands = {"Plains", "Island", "Swamp", "Mountain", "Forest"};
@@ -184,24 +204,35 @@ public class DeckServiceJpa implements DeckService {
             cardInDeckRepo.delete(card);
         }
 
+        Set<CardInDeck> cards = new HashSet<>();
+
         for (CardInDeckDTO card : main_deck) {
-            cardInDeckRepo.save(new CardInDeck(deck, card.getCard_name(), card.getQuantity(), true));
+            CardInDeck newCard = new CardInDeck(deck, card.getCard_name(), card.getQuantity(), true);
+            cardInDeckRepo.save(newCard);
+            cards.add(newCard);
         }
 
         for (CardInDeckDTO card : sideboard) {
-            cardInDeckRepo.save(new CardInDeck(deck, card.getCard_name(), card.getQuantity(), false));
+            CardInDeck newCard = new CardInDeck(deck, card.getCard_name(), card.getQuantity(), false);
+            cardInDeckRepo.save(newCard);
+            cards.add(newCard);
         }
 
-        Deck newDeck = deckRepo.findById(deck.getDeck_id()).get();
+        deck.setCards(cards);
 
-        return newDeck;
+        deckRepo.save(deck);
+
+        return deck;
     }
 
     @Override
-    public void deleteDeck(long id) {
+    public void deleteDeck(long id, String username) {
         Optional<Deck> deckCheck = deckRepo.findById(id);
         if (!deckCheck.isEmpty()) {
             Deck deck = deckCheck.get();
+            if (!username.equals(deck.getCreator().getUsername())) {
+                throw new IllegalArgumentException("You can't delete other player's decks.");
+            }
             for (CardInDeck card : deck.getCards()) {
                 cardInDeckRepo.delete(card);
             }
